Improve bottom tab switching by hiding PopoverMenu
